// green-store/app/lib/api/models/product-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Store Management API
 * API for managing store operations including inventory, orders, and customers
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ProductResponse {
    'productId'?: number;
    'productName'?: string | null;
    'categoryName'?: string | null;
    'supplierName'?: string | null;
    'barcode'?: string | null;
    'price'?: number;
    'unit'?: string | null;
    'imageUrl'?: string | null;
    'createdAt'?: string;
    'stockQuantity'?: number | null;
}